Update and rename Mtrix.ipynb to 2. Chain matrix multiplication.ipynb
Problem Statement: Matrix Chain Multiplication
You are given a sequence of matrices, and you need to determine the most efficient way to multiply these matrices together. The objective is to minimize the total number of scalar multiplications required to multiply the matrices in the given sequence.

Input
Number of Matrices: An integer n representing the number of matrices.

Matrix Dimensions: A list of n pairs of integers where each pair (r, c) represents the dimensions of a matrix. The matrix at position i has dimensions r x c.

The first matrix has dimensions dim[0][0] x dim[0][1].
The second matrix has dimensions dim[1][0] x dim[1][1], and so on.
The multiplication order for matrices A1, A2, ..., An will be determined based on these dimensions.
Output
Minimum Number of Scalar Multiplications: An integer representing the minimum number of scalar multiplications required to multiply the matrices together in the optimal order.
Optimal Matrix Multiplication Order: A string representing the optimal order of matrix multiplications using parentheses to indicate the grouping.